Bug 708464: Work around overflow in Freetype
On environments where longs are 4 bytes (such as Windows), one of the
calculations can overflow at higher resolutions/glyph sizes and, in those case,
with rotation involved.

As a workaround, promote the types from 4 byte "long" to 8 byte "long long", and
the same for the unsigned counterpart.

+++ freetype/src/raster/ftraster.c
@@ -297,8 +297,8 @@
   typedef unsigned int    UInt;
   typedef short           Short;
   typedef unsigned short  UShort, *PUShort;
-  typedef long            Long, *PLong;
-  typedef unsigned long   ULong;
+  typedef long long           Long, *PLong;
+  typedef unsigned long long   ULong;
 
   typedef unsigned char   Byte, *PByte;
   typedef char            Bool;
